requirements.txt
# Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Banco de Dados
supabase==2.28.3
python-dotenv==1.0.0
websockets>=14.0

# ML - Traducao e Embeddings
sentence-transformers>=3.1.0
transformers>=4.45.0
torch==2.11.0

# Validacao
pydantic>=2.13.2
pydantic-settings>=2.14.0

# Utilitarios
httpx==0.28.1, 
numpy==2.2.1
sentencepiece==0.2.1
sacremoses==0.1.1

# Autenticacao
PyJWT==2.10.1